Added Database Update to the Live Game (Main Menu).
Database will be updated on Main Menu and Start will only be able if Data was loaded.
If there is no Internet there should a short delay and then the Data will be loaded from the last JSON file.

We still get an Error:
Error: a native collection has not been disposed resulting in a memory leak